+++ Frontend/game-connect/src/components/ChatComponents/Online/index.js
@@ -1,9 +1,42 @@
 import './OnlineElements.css';
-import {useState, useEffect} from 'react';
+import {useState, useEffect, useContext} from 'react';
 import axios from 'axios';
+import {AuthContext} from '../../../context/AuthContext';
 
 const Online = ({onlineUsers, currentUser, setCurrentChat}) => {
-    
+    // Username
+    const {user} = useContext(AuthContext);
+    const [onlineFriends, setOnlineFriends] = useState([]);
+
+    // Get friend list from backend   
+    useEffect(() => {
+        const source = axios.CancelToken.source();
+
+        const getOnlineFriends = async (user) => {
+            try {
+                const res = await axios.get("backend/users/friends/" + user, {
+                    cancelToken: source.token,
+                });
+                var friendList = res.data.filter((each_friend) => onlineUsers.some((each_socket_user) => each_socket_user.userName === each_friend.username))
+
+                setOnlineFriends(friendList);
+            }   
+            catch (error) {
+                if (axios.isCancel(error)){
+
+                } else {
+                    console.log(error);
+                }
+            }
+        }
+
+        getOnlineFriends(user);
+
+        return () => {
+            source.cancel();
+        }
+    }, [user, onlineUsers]);
+
     const setConversation = async (user) => {
         try {
             
@@ -18,8 +51,8 @@ const Online = ({onlineUsers, currentUser, setCurrentChat}) => {
     return(
         <div className="online">
             {
-                onlineUsers.length !== 0 ?
-                    onlineUsers.map (each_online_friend => (
+                onlineFriends.length !== 0 ?
+                    onlineFriends.map (each_online_friend => (
                         <div key = {each_online_friend._id} className="onlineFriend" onClick = {() => setConversation (each_online_friend.username)}>
                             <div className="onlineImageContainer">
                                 <img className="onlineImage"
